Extract benchmark to own file
